Add files via upload
•	Summarize the project and what problem it was solving.
The code takes in a list of items from a text file it creates, giving the user three options to print the list's contents. From the first option, the user inputs an existing word on the list, and the system automatically capitalizes the first letter and then outputs that word, followed by the number of times it is repeated in the file. The second option outputs every word in the text file, followed by the number of repetitions. Option three is like option two. However, instead of a number printed after each word, a histogram of “*” represents the number's value.
•	What did you do particularly well?
I had been trying to figure out how to get a class’s functions to operate in the main function, and I succeeded in this project. I believe I kept trying to access them like standard functions without declaring the class at the beginning of the main. Utilizing this in the future will give me a better understanding of OOP, and now I can go back to previous projects and implement this more efficiently.
•	Where could you enhance your code? How would these improvements make your code more efficient, secure, and so on?
I could see improvement in my coding in a couple of ways. The first way is how I write/ create a text file. There must be an easier way to transfer or drag and drop the data into the source files. The next comes from how I needed to write all the variables repeatedly in each function. I need to work on setting up the case variables instead of repeating this each time. The last thing I need to work on to shorten the code is opening the file only once. I repeatedly open and close the file after each option choice in each function. Considering I do not know what adverse side effects are caused by this, I need to work on opening the file in the beginning and closing it at the end. 
•	Which pieces of the code did you find most challenging to write, and how did you overcome this? What tools or resources are you adding to your support network?
Two parts of this code were challenging to write. The first involved writing each line of text to the text file. As I mentioned above, there has got to be an easier way to do this. Considering it involved 102 lines of text, this was time-consuming and repetitive. The second portion was getting the class to operate. Considering this was the first time I got it to work in a C++ program, it wasn't easy to operate initially. I mainly focused on using W3 Schools as a source to help figure out the answers, but I have come to utilize C Plus Plus and YouTube tutorial channels to help figure out some of the more elaborate types of programming. This is because I am getting the human aspect of sharing ideas and thoughts I do not get from W3 Schools.
•	What skills from this project will be particularly transferable to other projects or coursework?
Some parts of this code that are transferable to other programs or ways that I can use it to improve previous projects are the writing to a text file portion. If this is the only way, then this shows how to do this. The second is OOP design, where I can expand on this to build more elaborate programs. The next is using a map, which I am still unsure how it operates because I have only read one chapter on this subject. Still, a code like this was handy, and understanding how it can be used can prove helpful in future projects.
•	How did you make this program maintainable, readable, and adaptable?
I am separating each portion of the code into blocks like main, class, and writing the text file. I can easily track down which portion has the error and go from there. If there is a problem with reading the file, I know that the file is written in the writing text file portion, so the problem is not there. Main calls for functions from class and never opens a file, so the issue is not in main. This leaves the class the only one of the three that opens and reads the file; therefore, class is where I need to look. The flow follows a standard program with different functions and classes ahead of the main. With in-line comments, other programmers unfamiliar with my code can easily follow along and figure out what it is designed to do. Also, the code can incorporate updates to the program by allowing new code to be written in the class section and then called in the main portion of the program.